# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TriLink Platform API
  description: |
    Comprehensive API documentation for the TriLink Platform - a multi-party procurement and logistics management system.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Roles & Permissions
    The platform supports the following roles:
    - **Buyer**: Creates purchase requests, evaluates bids, manages contracts
    - **Supplier**: Responds to RFQs, submits bids
    - **Logistics**: Manages shipments, updates GPS tracking
    - **Clearance**: Handles customs clearance
    - **Service Provider**: Provides various services
    - **Government**: View-only access with dispute resolution capabilities
    - **Admin**: Full system access
    
    ## Company Isolation
    Most resources are isolated by company. Users can only access resources belonging to their company unless they have Admin role.
  version: 1.0.0
  contact:
    name: TriLink Platform Support
    email: support@trilink.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.trilink.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Creates a new user account and returns access token and user data.
        Refresh token is set in httpOnly cookie (not in response body).
        Rate limited to prevent abuse.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePass123!
              role: Buyer
              companyId: 507f1f77bcf86cd799439011
              firstName: John
              lastName: Doe
              phone: +971501234567
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: Refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticates user and returns access token and user data.
        Refresh token is set in httpOnly cookie (not in response body).
        Rate limited to prevent brute force attacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Generates a new access token using a valid refresh token from httpOnly cookie.
        Implements token rotation for enhanced security.
        Refresh token is automatically sent via httpOnly cookie (not in request body).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              description: New refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: |
        Sends a password reset email with a secure token. The token expires in 1 hour.
        For security reasons, this endpoint always returns success even if the email doesn't exist,
        preventing email enumeration attacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: user@example.com
      responses:
        '200':
          description: Password reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: If an account exists with this email, a password reset link has been sent.
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: |
        Resets user password using a valid reset token from the forgot-password email.
        The token must be valid and not expired (1 hour expiry).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              password: NewSecurePass123!
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Password has been reset successfully.
                  requestId:
                    type: string
        '400':
          description: Invalid or expired token, or password validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Account is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: |
        Logs out the current user by clearing the refresh token httpOnly cookie.
        Access token should be discarded by the client.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              description: Refresh token cookie cleared
              schema:
                type: string
                example: refreshToken=; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=0
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Logged out successfully.
                  requestId:
                    type: string
...
  /bids:
    post:
      tags:
        - Bids
      summary: Create bid
      security:
        - bearerAuth: []
      description: Creates a new bid for an RFQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBidRequest'
      responses:
        '201':
          description: Bid created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBid'

    get:
      tags:
        - Bids
      summary: Get bids
      security:
        - bearerAuth: []
      description: Retrieves bids filtered by company
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BidResponse'
  /bids/rfq/{rfqId}:
    get:
      tags:
        - Bids
      summary: Get bids by RFQ
      security:
        - bearerAuth: []
      description: Retrieves all bids for a specific RFQ
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BidResponse'
  /bids/{id}:
    get:
      tags:
        - Bids
      summary: Get bid by ID
      security:
        - bearerAuth: []
      description: Retrieves a specific bid
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Bid details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBid'

    patch:
      tags:
        - Bids
      summary: Update bid
      security:
        - bearerAuth: []
      description: Updates a bid
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBidRequest'
      responses:
        '200':
          description: Bid updated successfully

    delete:
      tags:
        - Bids
      summary: Delete bid
      security:
        - bearerAuth: []
      description: Deletes a bid
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Bid deleted successfully
  /bids/{id}/withdraw:
    post:
      tags:
        - Bids
      summary: Withdraw bid
      security:
        - bearerAuth: []
      description: Withdraws a submitted bid
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Bid withdrawn successfully
  /bids/{id}/evaluate:
    post:
      tags:
        - Bids
      summary: Evaluate bid
      security:
        - bearerAuth: []
      description: Evaluates a bid (accept/reject). Requires EVALUATE_BID permission.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [accepted, rejected]
                notes:
                  type: string
      responses:
        '200':
          description: Bid evaluated successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        requestId:
          type: string
...
    Attachment:
      type: object
      properties:
        type:
          type: string
        url:
          type: string
        uploadedAt:
          type: string
          format: date-time

    # Authentication schemas
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - role
        - companyId
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [Buyer, Supplier, Logistics, Clearance, Service Provider, Government, Admin]
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    # RefreshTokenRequest removed - refresh token comes from httpOnly cookie, not request body
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address of the account to reset password for
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: Password reset token received via email
        password:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)

    # AuthResponse kept for backward compatibility but not used
    # Use LoginResponse instead which doesn't include refreshToken in body
...
    LoginResponse:
      type: object
      description: Response for login and register endpoints
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              description: JWT access token (15 minutes expiry)
          # refreshToken is NOT in response - set in httpOnly cookie
        requestId:
          type: string
    RefreshTokenResponse:
      type: object
      description: Response for refresh token endpoint
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              description: New JWT access token (15 minutes expiry)
          # refreshToken is NOT in response - new token set in httpOnly cookie
        requestId:
          type: string

    # TokenResponse removed - use RefreshTokenResponse instead

    # User schemas
    User:
      type: object
      description: User object returned in authentication responses
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Buyer, Supplier, Logistics, Clearance, Service Provider, Government, Admin]
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    CreateBidRequest:
      type: object
      required:
        - rfqId
        - price
        - paymentTerms
        - deliveryTime
        - deliveryDate
        - validity
      properties:
        rfqId:
          type: string
        price:
          type: number
        currency:
          type: string
          default: AED
        paymentTerms:
          type: string
        deliveryTime:
          type: number
          description: Delivery time in days
        deliveryDate:
          type: string
          format: date-time
        validity:
          type: string
          format: date-time
        anonymousBidder:
          type: boolean
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
    UpdateBidRequest:
      type: object
      properties:
        price:
          type: number
        paymentTerms:
          type: string
        deliveryTime:
          type: number
        deliveryDate:
          type: string
          format: date-time
        validity:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, submitted, under_review, accepted, rejected, withdrawn]
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
    BidResponse:
      type: object
      properties:
        _id:
          type: string
        rfqId:
          type: string
        companyId:
          type: string
        providerId:
          type: string
        price:
          type: number
        currency:
          type: string
        paymentTerms:
          type: string
        deliveryTime:
          type: number
        deliveryDate:
          type: string
          format: date-time
        validity:
          type: string
          format: date-time
        aiScore:
          type: number
        status:
          type: string
        anonymousBidder:
          type: boolean
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponseBid:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BidResponse'

    # Contract schemas
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'